List component versions for all three OSes
Use `bintool versions` to build the list.  Its output includes the commit
hashes of any overridden projects, but not in a form that GitHub will
linkify, and does not include the openslide-bin commit hash.  For
simplicity and for consistency with release builds, call this good enough,
and stop hand-modifying the version list.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>